Move met coordinator to own module
